ESDEV-3493 Refactored tests
Split test into a general test, which test the DatabaseInterfaceImplementation and a test for each Implementation (Doctrine and LegacyDatabase)
