+++ tests/integration/core/Database/LegacyDatabaseTest.php
@@ -1,0 +1,123 @@
+<?php
+namespace OxidEsales\Eshop\Tests\integration\core\Database;
+
+    /**
+     * This file is part of OXID eShop Community Edition.
+     *
+     * OXID eShop Community Edition is free software: you can redistribute it and/or modify
+     * it under the terms of the GNU General Public License as published by
+     * the Free Software Foundation, either version 3 of the License, or
+     * (at your option) any later version.
+     *
+     * OXID eShop Community Edition is distributed in the hope that it will be useful,
+     * but WITHOUT ANY WARRANTY; without even the implied warranty of
+     * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+     * GNU General Public License for more details.
+     *
+     * You should have received a copy of the GNU General Public License
+     * along with OXID eShop Community Edition.  If not, see <http://www.gnu.org/licenses/>.
+     *
+     * @link          http://www.oxid-esales.com
+     * @copyright (C) OXID eSales AG 2003-2016
+     * @version       OXID eShop CE
+     */
+
+
+/**
+ * Tests for our database object.
+ *
+ * @group doctrine
+ */
+class LegacyDatabaseTest extends DatabaseInterfaceImplementationTest
+{
+
+    const DATABASE_EXCEPTION_CLASS = 'oxAdoDbException';
+
+    const RESULT_SET_CLASS = 'ADORecordSet';
+
+    const EMPTY_RESULT_SET_CLASS = 'ADORecordSet_empty';
+
+    /**
+     * Set up before beginning with tests
+     */
+    public static function setUpBeforeClass()
+    {
+        parent::setUpBeforeClass();
+
+        self::createDatabaseTable();
+    }
+
+    /**
+     * Tear down after all tests are done
+     */
+    public static function tearDownAfterClass()
+    {
+        self::removeDatabaseTable();
+
+        parent::tearDownAfterClass();
+    }
+
+    /**
+     * Create a table in the database especially for this test.
+     */
+    protected static function createDatabaseTable()
+    {
+        $db = self::createDatabaseStatic();
+
+        $db->execute('CREATE TABLE IF NOT EXISTS ' . self::TABLE_NAME . ' (oxid CHAR(32), oxuserid CHAR(32)) ENGINE innoDb;');
+    }
+
+    /**
+     * Drop the test database table.
+     */
+    protected static function removeDatabaseTable()
+    {
+        $db = self::createDatabaseStatic();
+
+        $db->execute('DROP TABLE ' . self::TABLE_NAME . ';');
+    }
+
+    /**
+     * @return string The name of the database exception class
+     */
+    protected function getDatabaseExceptionClassName()
+    {
+        return self::DATABASE_EXCEPTION_CLASS;
+    }
+
+    /**
+     * @return string The name of the result set class
+     */
+    protected function getResultSetClassName()
+    {
+        return self::RESULT_SET_CLASS;
+    }
+
+    /**
+     * @return string The name of the empty result set class
+     */
+    protected function getEmptyResultSetClassName()
+    {
+        return self::EMPTY_RESULT_SET_CLASS;
+    }
+
+    /**
+     * Create the database object under test.
+     *
+     * @return \oxLegacyDb The database object under test.
+     */
+    protected function createDatabase()
+    {
+        return \oxDb::getDb();
+    }
+
+    /**
+     * Create the database object under test - the static pendant to use in the setUpBeforeClass and tearDownAfterClass.
+     *
+     * @return \oxLegacyDb The database object under test.
+     */
+    protected static function createDatabaseStatic()
+    {
+        return \oxDb::getDb();
+    }
+}
